Experimenting with a method for automatically setting fit vars.
Trying a pass-through scheme where all the plotting functions of
mpl.axis.Axes objects will be re-implemented as a method of the
QMPLFitterWidget with the same exact method name. Then the user can
just implement parsing functionality to pull the relevant fitting
info from the inputs and pass the info along to the plotting.