Modified example_standalone to work when called as script.
example_standalone originally required that QApplication already be
running for the QMPLWidgets to work.
Thus it worked when called in an ipython terminal where matplotlib with
the appropriate backend was already imported, but didn't work when
called as a script (python example_standalone.py).
Now works with method for running the example.