Added files to lower barrier-to-entry.
requirements.txt to catalog dependencies of project.
N.B. - build fails with PySide==1.2.1, so updated to 1.2.4.

Currently working only with the Qt4Agg backend, so include a
matplotlibrc to the project to force the use of this backend
by default.

N.B. - Users may also have to set the QT_API env var to force
the Qt4Agg backend to use PySide bindings (depending on MPL version)
Should try to remove this dependency and allow for PyQt4 as well.